version 2.7.30
[Bugfix]
	Fixed bug FloatType to DoubleType @productdailysalessummarydao

+++ ever-grocery/src/main/java/com/chua/evergrocery/database/dao/impl/ProductDailySalesSummaryDAOImpl.java
@@ -10,7 +10,7 @@ import org.hibernate.criterion.Projections;
 import org.hibernate.criterion.Restrictions;
 import org.hibernate.sql.JoinType;
 import org.hibernate.transform.Transformers;
-import org.hibernate.type.FloatType;
+import org.hibernate.type.DoubleType;
 import org.springframework.stereotype.Repository;
 
 import com.chua.evergrocery.beans.ProductSalesSummaryBean;
@@ -57,8 +57,8 @@ public class ProductDailySalesSummaryDAOImpl
 		
 		final ProjectionList pList = Projections.projectionList();
 		pList.add(Projections.groupProperty("p.id"), "productId");
-		pList.add(Projections.sqlProjection("sum(net_total) as netTotal", new String[] { "netTotal" }, new FloatType[] { new FloatType() }), "netTotal");
-		pList.add(Projections.sqlProjection("sum(base_total) as baseTotal", new String[] { "baseTotal" }, new FloatType[] { new FloatType() }), "baseTotal");
+		pList.add(Projections.sqlProjection("sum(net_total) as netTotal", new String[] { "netTotal" }, new DoubleType[] { new DoubleType() }), "netTotal");
+		pList.add(Projections.sqlProjection("sum(base_total) as baseTotal", new String[] { "baseTotal" }, new DoubleType[] { new DoubleType() }), "baseTotal");
 		
 		String[] associatedPaths = { "product" };
 		String[] aliasNames = { "p" };
